docs(core): overflow layout stackblitz
